src/main/webapp/app/entities/day/update/day-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { DayService } from '../service/day.service';
import { IDay, Day } from '../day.model';

import { DayUpdateComponent } from './day-update.component';

describe('Day Management Update Component', () => {
  let comp: DayUpdateComponent;
  let fixture: ComponentFixture<DayUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let dayService: DayService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [DayUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(DayUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(DayUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    dayService = TestBed.inject(DayService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const day: IDay = { id: 456 };

      activatedRoute.data = of({ day });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(day));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Day>>();
      const day = { id: 123 };
      jest.spyOn(dayService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ day });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: day }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(dayService.update).toHaveBeenCalledWith(day);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Day>>();
      const day = new Day();
      jest.spyOn(dayService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ day });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: day }));
      saveSubject.complete();

      // THEN
      expect(dayService.create).toHaveBeenCalledWith(day);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Day>>();
      const day = { id: 123 };
      jest.spyOn(dayService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ day });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(dayService.update).toHaveBeenCalledWith(day);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
